Combine "LoRaDeviceClientConnectionManager" parts
